Seperated the logic for the navigation bars
