+++ composeApp/src/commonMain/kotlin/me/androidbox/busbymoviesv2/move_list/presentation/screens/BottomNavigationItems.kt
@@ -14,7 +14,7 @@ import me.androidbox.busbymoviesv2.move_list.presentation.MovieCategories
 import org.jetbrains.compose.resources.StringResource
 
 
-data class BottomMovieListNavigationItem(
+data class MovieListNavigationItem(
     val title: StringResource,
     val movieCategory: MovieCategories,
     val selectedIcon: ImageVector,
@@ -24,14 +24,14 @@ data class BottomMovieListNavigationItem(
 )
 
 val listOfNavigationItems = listOf(
-    BottomMovieListNavigationItem(
+    MovieListNavigationItem(
         title = MovieCategories.NOW_PLAYING.titleRes,
         movieCategory = MovieCategories.NOW_PLAYING,
         selectedIcon = Icons.Filled.PlayArrow,
         unSelectedIcon = Icons.Outlined.PlayArrow,
         hasExtra = false
     ),
-    BottomMovieListNavigationItem(
+    MovieListNavigationItem(
         title = MovieCategories.TOP_RATED.titleRes,
         movieCategory = MovieCategories.TOP_RATED,
         selectedIcon = Icons.Filled.Star,
@@ -39,14 +39,14 @@ val listOfNavigationItems = listOf(
         hasExtra = false,
         badgeCount = 5
     ),
-    BottomMovieListNavigationItem(
+    MovieListNavigationItem(
         title = MovieCategories.POPULAR.titleRes,
         movieCategory = MovieCategories.POPULAR,
         selectedIcon = Icons.Filled.ThumbUp,
         unSelectedIcon = Icons.Outlined.ThumbUp,
         hasExtra = false
     ),
-    BottomMovieListNavigationItem(
+    MovieListNavigationItem(
         title = MovieCategories.UPCOMING.titleRes,
         movieCategory = MovieCategories.UPCOMING,
         selectedIcon = Icons.Filled.DateRange,
